fix(seeder): to have deterministric generation we need to generate
a brand new seeder on each calls, otherwise each users will the same
seed.

It can also create a race condition if multiple users get a specific
hero, as the seed is a package variable.

With this change, we put a new seeder in request context, avoiding race
and reusing the same seed over and over.